github.com/vbo/ms/requirements.txt
Flask<1
botbuilder-core>=4.4.0b1
botbuilder-dialogs>=4.4.0b1
aiohttp